+++ lib/mix/tasks/pleroma/database.ex
@@ -122,6 +122,37 @@ defmodule Mix.Tasks.Pleroma.Database do
     |> Stream.run()
   end
 
+  def run(["fill_old_hashtags"]) do
+    import Ecto.Query
+
+    start_pleroma()
+
+    from(
+      o in Object,
+      where: fragment("(?)->>'hashtags' is null", o.data),
+      where: fragment("(?)->>'tag' != '[]'", o.data),
+      select: %{id: o.id, tag: fragment("(?)->>'tag'", o.data)}
+    )
+    |> Pleroma.RepoStreamer.chunk_stream(100)
+    |> Stream.each(fn objects ->
+      objects
+      |> Enum.map(fn object ->
+        tags =
+          object.tag
+          |> Jason.decode!()
+          |> Enum.filter(&is_bitstring(&1))
+
+        Object
+        |> where([o], o.id == ^object.id)
+        |> update([o],
+          set: [data: fragment("safe_jsonb_set(?, '{hashtags}', ?, true)", o.data, ^tags)]
+        )
+        |> Repo.update_all([], timeout: :infinity)
+      end)
+    end)
+    |> Stream.run()
+  end
+
   def run(["vacuum", args]) do
     start_pleroma()