+++ test/pleroma/web/activity_pub/transmogrifier_test.exs
@@ -41,7 +41,7 @@ defmodule Pleroma.Web.ActivityPub.TransmogrifierTest do
       {:ok, %Activity{data: data, local: false}} = Transmogrifier.handle_incoming(data)
       object = Object.normalize(data["object"])
 
-      assert "test" in object.data["tag"]
+      assert "test" in object.data["hashtags"]
     end
 
     test "it cleans up incoming notices which are not really DMs" do
@@ -212,7 +212,7 @@ defmodule Pleroma.Web.ActivityPub.TransmogrifierTest do
       {:ok, %Activity{data: data, local: false}} = Transmogrifier.handle_incoming(data)
       object = Object.normalize(data["object"])
 
-      assert Enum.at(object.data["tag"], 2) == "moo"
+      assert object.data["hashtags"] == ["moo"]
     end
 
     test "it works for incoming notices with contentMap" do